+++ vnav_acquisition/config.py
@@ -3,10 +3,12 @@ import json
 
 class Configurator:
     _DEFAULT_CONFIG = {
-        'connection': ["raspberrypi", 22, "pi", "VibroNav"],
+        'connection': ["192.168.1.100", 22, "pi", "VibroNav"],
         'materials': ["Slime", "Silicone", "PU", "Plato (play dough)", "Plastic", "Ikea (plastic bag)",
                       "African (silk)"],
         'speeds': ["slow", "medium", "fast"],
+        'needleTypes': ["Type1", "Type2", "Type3"],
+        'microphoneTypes': ["TypeA", "TypeB", "TypeC"],
         'local_dir': "micro_data",
         'remote_dir': "vnav_acquisition"
     }